fix(dashboard): server-load meeting history
